backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
pydantic==2.9.2
python-dotenv==1.0.1
google-generativeai==0.8.3
selenium==4.24.0
webdriver-manager==4.0.2